Make importer class idempotent
As we are exporting the CSV data from the chat application in it's entirety and
not using snapshots, and as we want the dashboard to be in-sync with the chat
application on completion of the export/import process. We want the importer that
imports the exported CSV data to be idempotent.

That is, no matter how many times we run the importer - we want the exact result each
time. In this case, each time we should end up with the same number of records
that have been exported to be imported for both the Chat and Feedback models, have
the correct number of matching Answer records created, and have the same primary and
foreign keys created so that the model realtionships remain intact.

Only when the latest CSV export includes additional data, should the numbers of
records change and increase accordingly.

For our purposes we can do this simply by deleting the existing records - that we
no longer need as they will be replicated, then resetting the primary key sequence
on those specific tables.